// client/src/js/bindings/IntrospectResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Game } from "./Game";
import type { Permission } from "./Permission";

export type IntrospectResponse = { gamertag: string, game: Game, cert_not_after: bigint | null, permissions: Array<Permission>, };